requirements.txt
Flask
openai
firebase-admin
python-dotenv
